// src/generated/models/Axm365_eventsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Axm365_eventsaxm365_eventtype = {
  216260000: 'Training',
  216260001: 'TournamentMatch',
  216260002: 'LeagueMatch',
  216260003: 'FriendlyMatch'
} as const;
export type Axm365_eventsaxm365_eventtype = keyof typeof Axm365_eventsaxm365_eventtype;
export const Axm365_eventsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Axm365_eventsstatecode = keyof typeof Axm365_eventsstatecode;
export const Axm365_eventsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Axm365_eventsstatuscode = keyof typeof Axm365_eventsstatuscode;

export interface Axm365_eventsBase {
  "axm365_Coach@odata.bind"?: string;
  axm365_description?: string;
  axm365_eventdate?: string;
  axm365_eventid: string;
  axm365_eventtype?: Axm365_eventsaxm365_eventtype;
  "axm365_Facility@odata.bind"?: string;
  axm365_name: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Axm365_eventsstatecode;
  statuscode?: Axm365_eventsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Axm365_events extends Axm365_eventsBase {
  axm365_coachname?: string;
  axm365_eventtypename?: string;
  axm365_facilityname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  axm365_coach?: object;
  _axm365_coach_value?: string;
  axm365_facility?: object;
  _axm365_facility_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}