// src/generated/models/Axm365_eventcoachattendancesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Axm365_eventcoachattendancesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Axm365_eventcoachattendancesstatecode = keyof typeof Axm365_eventcoachattendancesstatecode;
export const Axm365_eventcoachattendancesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Axm365_eventcoachattendancesstatuscode = keyof typeof Axm365_eventcoachattendancesstatuscode;

export interface Axm365_eventcoachattendancesBase {
  axm365_clockin?: string;
  axm365_clockout?: string;
  "axm365_Coach@odata.bind"?: string;
  "axm365_Event@odata.bind"?: string;
  axm365_eventcoachattendanceid: string;
  axm365_name: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Axm365_eventcoachattendancesstatecode;
  statuscode?: Axm365_eventcoachattendancesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Axm365_eventcoachattendances extends Axm365_eventcoachattendancesBase {
  axm365_coachname?: string;
  axm365_duration?: number;
  axm365_eventname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  axm365_coach?: object;
  _axm365_coach_value?: string;
  axm365_event?: object;
  _axm365_event_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}